requirements.txt
streamlit 
pymupdf 
python-docx 
pandas 
openpyxl
pdfplumber